@@ -7,167 +7,361 @@ namespace SurfTimer;
 
 public partial class SurfTimer
 {
-    // Trigger start touch handler - CBaseTrigger_StartTouchFunc
-    // internal HookResult OnTriggerStartTouch(DynamicHook handler)
+    /// <summary>
+    /// Handler for trigger start touch hook - CBaseTrigger_StartTouchFunc
+    /// </summary>
+    /// <returns>CounterStrikeSharp.API.Core.HookResult</returns>
+    /// <exception cref="Exception"></exception>
     internal HookResult OnTriggerStartTouch(CEntityIOOutput output, string name, CEntityInstance activator, CEntityInstance caller, CVariant value, float delay)
     {
-        // CBaseTrigger trigger = handler.GetParam<CBaseTrigger>(0);
         CBaseTrigger trigger = new CBaseTrigger(caller.Handle);
-        // CBaseEntity entity = handler.GetParam<CBaseEntity>(1);
         CBaseEntity entity = new CBaseEntity(activator.Handle);
         CCSPlayerController client = new CCSPlayerController(new CCSPlayerPawn(entity.Handle).Controller.Value!.Handle);
         if (!client.IsValid || !client.PawnIsAlive || !playerList.ContainsKey((int)client.UserId!)) // !playerList.ContainsKey((int)client.UserId!) make sure to not check for user_id that doesnt exists
         {
             return HookResult.Continue;
         }
-        else
+        // To-do: Sometimes this triggers before `OnPlayerConnect` and `playerList` does not contain the player how is this possible :thonk:
+        if (!playerList.ContainsKey(client.UserId ?? 0))
         {
-            // To-do: Sometimes this triggers before `OnPlayerConnect` and `playerList` does not contain the player how is this possible :thonk:
-            if (!playerList.ContainsKey(client.UserId ?? 0))
-            {
-                Console.WriteLine($"CS2 Surf ERROR >> OnTriggerStartTouch -> Init -> Player playerList does NOT contain client.UserId, this shouldn't happen. Player: {client.PlayerName} ({client.UserId})");
-                throw new Exception($"CS2 Surf ERROR >> OnTriggerStartTouch -> Init -> Player playerList does NOT contain client.UserId, this shouldn't happen. Player: {client.PlayerName} ({client.UserId})");
-                // return HookResult.Continue;
-            }
-            // Implement Trigger Start Touch Here
-            Player player = playerList[client.UserId ?? 0];
-            #if DEBUG
-            player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc -> {trigger.DesignerName} -> {trigger.Entity!.Name}");
-            #endif
+            Console.WriteLine($"CS2 Surf ERROR >> OnTriggerStartTouch -> Init -> Player playerList does NOT contain client.UserId, this shouldn't happen. Player: {client.PlayerName} ({client.UserId})");
+            Exception exception = new($"CS2 Surf ERROR >> OnTriggerStartTouch -> Init -> Player playerList does NOT contain client.UserId, this shouldn't happen. Player: {client.PlayerName} ({client.UserId})");
+            throw exception;
+        }
+        // Implement Trigger Start Touch Here
+        Player player = playerList[client.UserId ?? 0];
+#if DEBUG
+        player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc -> {trigger.DesignerName} -> {trigger.Entity!.Name}");
+#endif
 
-            if (trigger.Entity!.Name != null)
-            {
-                // Get velocities for DB queries
-                // Get the velocity of the player - we will be using this values to compare and write to DB
-                float velocity_x = player.Controller.PlayerPawn.Value!.AbsVelocity.X;
-                float velocity_y = player.Controller.PlayerPawn.Value!.AbsVelocity.Y;
-                float velocity_z = player.Controller.PlayerPawn.Value!.AbsVelocity.Z;
-                float velocity = (float)Math.Sqrt(velocity_x * velocity_x + velocity_y * velocity_y + velocity_z + velocity_z);
-                int style = player.Timer.Style;
+        if (DB == null)
+        {
+            Exception exception = new Exception("CS2 Surf ERROR >> OnTriggerStartTouch (Map end zone) -> DB object is null, this shouldn't happen.");
+            throw exception;
+        }
 
-                // Map end zones -- hook into map_end
-                if (trigger.Entity.Name == "map_end")
+        if (trigger.Entity!.Name != null)
+        {
+            // Get velocities for DB queries
+            // Get the velocity of the player - we will be using this values to compare and write to DB
+            float velocity_x = player.Controller.PlayerPawn.Value!.AbsVelocity.X;
+            float velocity_y = player.Controller.PlayerPawn.Value!.AbsVelocity.Y;
+            float velocity_z = player.Controller.PlayerPawn.Value!.AbsVelocity.Z;
+            int pStyle = player.Timer.Style;
+
+            // Map end zones -- hook into map_end
+            if (trigger.Entity.Name == "map_end")
+            {
+                player.Controller.PrintToCenter($"Map End");
+
+                player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.END_ZONE_ENTER;
+                player.ReplayRecorder.MapSituations.Add(player.Timer.Ticks);
+
+                player.Stats.ThisRun.Ticks = player.Timer.Ticks; // End time for the Map run
+                player.Stats.ThisRun.EndVelX = velocity_x; // End speed for the Map run
+                player.Stats.ThisRun.EndVelY = velocity_y; // End speed for the Map run
+                player.Stats.ThisRun.EndVelZ = velocity_z; // End speed for the Map run
+
+
+                // MAP END ZONE - Map RUN
+                if (player.Timer.IsRunning && !player.Timer.IsStageMode)
                 {
-                    player.Controller.PrintToCenter($"Map End");
-                    // MAP END ZONE
-                    if (player.Timer.IsRunning)
+                    player.Timer.Stop();
+                    bool saveMapTime = false;
+                    string PracticeString = "";
+                    if (player.Timer.IsPracticeMode)
+                        PracticeString = $"({ChatColors.Grey}Practice{ChatColors.Default}) ";
+
+                    if (player.Timer.Ticks < CurrentMap.WR[pStyle].Ticks) // Player beat the Map WR
                     {
-                        player.Timer.Stop();
-                        player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.END_RUN;
+                        saveMapTime = true;
+                        int timeImprove = CurrentMap.WR[pStyle].Ticks - player.Timer.Ticks;
+                        Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} has set a new {ChatColors.Yellow}Map{ChatColors.Default} record with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, beating the old record by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(CurrentMap.WR[pStyle].Ticks)}{ChatColors.Default})");
+                    }
+                    else if (CurrentMap.WR[pStyle].ID == -1) // No record was set on the map
+                    {
+                        saveMapTime = true;
+                        Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} set the first {ChatColors.Yellow}Map{ChatColors.Default} record at {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}!");
+                    }
+                    else if (player.Stats.PB[pStyle].Ticks <= 0) // Player first ever PersonalBest for the map
+                    {
+                        saveMapTime = true;
+                        player.Controller.PrintToChat($"{Config.PluginPrefix} {PracticeString}You finished the {ChatColors.Yellow}Map{ChatColors.Default} in {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, setting your new Personal Best!");
+                    }
+                    else if (player.Timer.Ticks < player.Stats.PB[pStyle].Ticks) // Player beating their existing PersonalBest for the map
+                    {
+                        saveMapTime = true;
+                        int timeImprove = player.Stats.PB[pStyle].Ticks - player.Timer.Ticks;
+                        Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} beat their {ChatColors.Yellow}Map{ChatColors.Default} Personal Best with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, improving by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(player.Stats.PB[pStyle].Ticks)}{ChatColors.Default})");
+                    }
+                    else // Player did not beat their existing PersonalBest for the map nor the map record
+                    {
+                        player.Controller.PrintToChat($"{Config.PluginPrefix} {PracticeString}You finished the map in {ChatColors.Yellow}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}!");
+                    }
 
-                        player.Stats.ThisRun.Ticks = player.Timer.Ticks; // End time for the run
-                        player.Stats.ThisRun.EndVelX = velocity_x; // End pre speed for the run
-                        player.Stats.ThisRun.EndVelY = velocity_y; // End pre speed for the run
-                        player.Stats.ThisRun.EndVelZ = velocity_z; // End pre speed for the run
-
-                        string PracticeString = "";
-                        if (player.Timer.IsPracticeMode)
-                            PracticeString = $"({ChatColors.Grey}Practice{ChatColors.Default}) ";
-
-                        // To-do: make Style (currently 0) be dynamic
-                        if (player.Stats.PB[style].Ticks <= 0) // Player first ever PersonalBest for the map
+                    if (saveMapTime)
+                    {
+                        player.ReplayRecorder.IsSaving = true;
+                        AddTimer(1.0f, async () =>
                         {
-                            Server.PrintToChatAll($"{PluginPrefix} {PracticeString}{player.Controller.PlayerName} finished the map in {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default} ({player.Timer.Ticks})!");
-                        }
-                        else if (player.Timer.Ticks < player.Stats.PB[style].Ticks) // Player beating their existing PersonalBest for the map
-                        {
-                            Server.PrintToChatAll($"{PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Profile.Name}{ChatColors.Default} beat their PB in {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default} (Old: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(player.Stats.PB[style].Ticks)}{ChatColors.Default})!");
-                        }
-                        else // Player did not beat their existing PersonalBest for the map
-                        {
-                            player.Controller.PrintToChat($"{PluginPrefix} {PracticeString}You finished the map in {ChatColors.Yellow}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}!");
-                            return HookResult.Continue; // Exit here so we don't write to DB
-                        }
+                            await player.Stats.ThisRun.SaveMapTime(player); // Save the MapTime PB data
+                        });
+                    }
 
-                        if (DB == null)
-                            throw new Exception("CS2 Surf ERROR >> OnTriggerStartTouch (Map end zone) -> DB object is null, this shouldn't happen.");
-
-                        
-                        player.Stats.PB[style].Ticks = player.Timer.Ticks; // Reload the run_time for the HUD and also assign for the DB query
-
-                        #if DEBUG
-                        Console.WriteLine($@"CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> 
+#if DEBUG
+                    Console.WriteLine($@"CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> 
                             ============== INSERT INTO `MapTimes` 
                             (`player_id`, `map_id`, `style`, `type`, `stage`, `run_time`, `start_vel_x`, `start_vel_y`, `start_vel_z`, `end_vel_x`, `end_vel_y`, `end_vel_z`, `run_date`) 
-                            VALUES ({player.Profile.ID}, {CurrentMap.ID}, {style}, 0, 0, {player.Stats.ThisRun.Ticks}, 
+                            VALUES ({player.Profile.ID}, {CurrentMap.ID}, {pStyle}, 0, 0, {player.Stats.ThisRun.Ticks}, 
                             {player.Stats.ThisRun.StartVelX}, {player.Stats.ThisRun.StartVelY}, {player.Stats.ThisRun.StartVelZ}, {velocity_x}, {velocity_y}, {velocity_z}, {(int)DateTimeOffset.UtcNow.ToUnixTimeSeconds()})
                             ON DUPLICATE KEY UPDATE run_time=VALUES(run_time), start_vel_x=VALUES(start_vel_x), start_vel_y=VALUES(start_vel_y), 
                             start_vel_z=VALUES(start_vel_z), end_vel_x=VALUES(end_vel_x), end_vel_y=VALUES(end_vel_y), end_vel_z=VALUES(end_vel_z), run_date=VALUES(run_date);
                         ");
-                        #endif
+#endif
 
-                        // Add entry in DB for the run
-                        if(!player.Timer.IsPracticeMode) {
-                            AddTimer(1.5f, async () => {
-                                player.Stats.ThisRun.SaveMapTime(player, DB); // Save the MapTime PB data
-                                player.Stats.LoadMapTimesData(player, DB); // Load the MapTime PB data again (will refresh the MapTime ID for the Checkpoints query)
-                                await player.Stats.ThisRun.SaveCurrentRunCheckpoints(player, DB); // Save this run's checkpoints
-                                player.Stats.LoadCheckpointsData(DB); // Reload checkpoints for the run - we should really have this in `SaveMapTime` as well but we don't re-load PB data inside there so we need to do it here
-                                CurrentMap.GetMapRecordAndTotals(DB); // Reload the Map record and totals for the HUD
-                            });
-
-                            // This section checks if the PB is better than WR
-                            if(player.Timer.Ticks < CurrentMap.WR[player.Timer.Style].Ticks || CurrentMap.WR[player.Timer.Style].ID == -1)
+                    // Add entry in DB for the run
+                    if (!player.Timer.IsPracticeMode)
+                    {
+                        // Should we also save a last stage run?
+                        if (CurrentMap.Stages > 0)
+                        {
+                            AddTimer(0.1f, () =>
                             {
-                                int WrIndex = CurrentMap.ReplayBots.Count-1; // As the ReplaysBot is set, WR Index will always be at the end of the List
-                                AddTimer(2f, () => {
-                                    CurrentMap.ReplayBots[WrIndex].Stat_MapTimeID = CurrentMap.WR[player.Timer.Style].ID;
-                                    CurrentMap.ReplayBots[WrIndex].LoadReplayData(DB!);
-                                    CurrentMap.ReplayBots[WrIndex].ResetReplay();
-                                });
-                            }
+                                // This calculation is wrong unless we wait for a bit in order for the `END_ZONE_ENTER` to be available in the `Frames` object
+                                int stage_run_time = player.ReplayRecorder.Frames.FindLastIndex(f => f.Situation == ReplayFrameSituation.END_ZONE_ENTER) - player.ReplayRecorder.Frames.FindLastIndex(f => f.Situation == ReplayFrameSituation.STAGE_ZONE_EXIT);
+
+                                // player.Controller.PrintToChat($"{Config.PluginPrefix} [LAST StageWR (Map RUN)] Sending to SaveStageTime: {player.Profile.Name}, {CurrentMap.Stages}, {stage_run_time}");
+                                SaveStageTime(player, CurrentMap.Stages, stage_run_time, true);
+                            });
                         }
+
+                        // This section checks if the PB is better than WR
+                        if (player.Timer.Ticks < CurrentMap.WR[pStyle].Ticks || CurrentMap.WR[pStyle].ID == -1)
+                        {
+                            AddTimer(2f, () =>
+                            {
+                                Console.WriteLine("CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> WR/PB");
+                                CurrentMap.ReplayManager.MapWR.Start(); // Start the replay again
+                                CurrentMap.ReplayManager.MapWR.FormatBotName();
+                            });
+                        }
+
                     }
 
-                    #if DEBUG
-                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Red}Map Stop Zone");
-                    #endif
+                    // API
+                    /*
+                    // Add entry in DB for the run
+                    if (!player.Timer.IsPracticeMode) {
+                        API_CurrentRun? last_stage_time = null;
+                        if (CurrentMap.Stages > 0)
+                        {
+                            int last_exit_tick = player.ReplayRecorder.LastExitTick();
+                            int last_enter_tick = player.ReplayRecorder.LastEnterTick();
+
+                            int stage_run_time = player.ReplayRecorder.Frames.Count - 1 - last_exit_tick; // Would like some check on this
+                            int time_since_last_enter = player.ReplayRecorder.Frames.Count - 1 - last_enter_tick;
+
+                            int tt = -1;
+                            if (last_exit_tick - last_enter_tick > 2*64)
+                                tt = last_exit_tick - 2*64;
+                            else
+                                tt = last_enter_tick;
+
+                            last_stage_time = new API_CurrentRun
+                            {
+                                    player_id = player.Profile.ID,
+                                    map_id = player.CurrMap.ID,
+                                    style = style,
+                                    type = 2,
+                                    stage = CurrentMap.Stages,
+                                    run_time = stage_run_time,
+                                    run_date = (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
+                                    replay_frames = player.ReplayRecorder.SerializeReplayPortion(tt, time_since_last_enter)
+                            };
+                        }
+                        AddTimer(1.5f, () => {
+                            List<API_Checkpoint> checkpoints = new List<API_Checkpoint>();
+                            foreach (var cp in player.Stats.ThisRun.Checkpoint)
+                            {
+                                checkpoints.Add(new API_Checkpoint
+                                {
+                                    cp = cp.Key,
+                                    run_time = cp.Value.Ticks,
+                                    start_vel_x = cp.Value.StartVelX,
+                                    start_vel_y = cp.Value.StartVelY,
+                                    start_vel_z = cp.Value.StartVelZ,
+                                    end_vel_x = cp.Value.EndVelX,
+                                    end_vel_y = cp.Value.EndVelY,
+                                    end_vel_z = cp.Value.EndVelZ,
+                                    end_touch = 0, // ?????
+                                    attempts = cp.Value.Attempts
+                                });
+                            }
+
+                            API_CurrentRun map_time = new API_CurrentRun
+                            {
+                                player_id = player.Profile.ID,
+                                map_id = player.CurrMap.ID,
+                                style = style,
+                                type = 0,
+                                stage = 0,
+                                run_time = player.Stats.ThisRun.Ticks,
+                                run_date = (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
+                                checkpoints = checkpoints,
+                                replay_frames = player.ReplayRecorder.SerializeReplay()
+                            };                                    
+
+                            Task.Run(async () => {
+                                System.Console.WriteLine("CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> Saved map time");
+                                await ApiCall.POST("/surftimer/savemaptime", map_time);
+
+                                if (last_stage_time != null)
+                                {
+                                    await ApiCall.POST("/surftimer/savestagetime", last_stage_time);
+                                    System.Console.WriteLine("CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> Saved last stage time");
+                                    player.Stats.LoadStageTime(player);
+                                }
+
+                                player.Stats.LoadMapTime(player);
+                                await CurrentMap.ApiGetMapRecordAndTotals(); // Reload the Map record and totals for the HUD
+                            });
+                        });
+
+                        // This section checks if the PB is better than WR
+                        if(player.Timer.Ticks < CurrentMap.WR[pStyle].Ticks || CurrentMap.WR[pStyle].ID == -1)
+                        {
+                            AddTimer(2f, () => {
+                                System.Console.WriteLine("CS2 Surf DEBUG >> OnTriggerStartTouch (Map end zone) -> WR PB");
+                                CurrentMap.ReplayManager.MapWR.LoadReplayData();
+
+                                AddTimer(1.5f, () => {
+                                    CurrentMap.ReplayManager.MapWR.FormatBotName();
+                                });
+                            });
+                        }
+                    }
+                    */
+                }
+                else if (player.Timer.IsStageMode)
+                {
+                    player.Timer.Stop();
+
+                    if (!player.Timer.IsPracticeMode)
+                    {
+                        AddTimer(0.1f, () =>
+                        {
+                            // This calculation is wrong unless we wait for a bit in order for the `END_ZONE_ENTER` to be available in the `Frames` object
+                            int stage_run_time = player.ReplayRecorder.Frames.FindLastIndex(f => f.Situation == ReplayFrameSituation.END_ZONE_ENTER) - player.ReplayRecorder.Frames.FindLastIndex(f => f.Situation == ReplayFrameSituation.STAGE_ZONE_EXIT);
+
+                            // player.Controller.PrintToChat($"{Config.PluginPrefix} [LAST StageWR (IsStageMode)] Sending to SaveStageTime: {player.Profile.Name}, {CurrentMap.Stages}, {stage_run_time}");
+                            SaveStageTime(player, CurrentMap.Stages, stage_run_time, true);
+                        });
+                    }
                 }
 
-                // Map start zones -- hook into map_start, (s)tage1_start
-                else if (trigger.Entity.Name.Contains("map_start") ||
-                        trigger.Entity.Name.Contains("s1_start") ||
-                        trigger.Entity.Name.Contains("stage1_start"))
-                {
-                    player.ReplayRecorder.Start(); // Start replay recording
+#if DEBUG
+                player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Red}Map Stop Zone");
+#endif
+            }
 
+            // Map start zones -- hook into map_start, (s)tage1_start
+            else if (trigger.Entity.Name.Contains("map_start") ||
+                    trigger.Entity.Name.Contains("s1_start") ||
+                    trigger.Entity.Name.Contains("stage1_start")
+            )
+            {
+                // We shouldn't start timer and reset data until MapTime has been saved - mostly concerns the Replays and trimming the correct parts
+                if (!player.ReplayRecorder.IsSaving)
+                {
+                    player.ReplayRecorder.Reset(); // Start replay recording
+                    player.ReplayRecorder.Start(); // Start replay recording
+                    player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.START_ZONE_ENTER;
+
+                    player.ReplayRecorder.MapSituations.Add(player.ReplayRecorder.Frames.Count);
+                    Console.WriteLine($"START_ZONE_ENTER: player.ReplayRecorder.MapSituations.Add({player.ReplayRecorder.Frames.Count})");
                     player.Timer.Reset();
-                    player.Stats.ThisRun.Checkpoint.Clear(); // I have the suspicion that the `Timer.Reset()` does not properly reset this object :thonk:
+                    player.Stats.ThisRun.Checkpoints.Clear();
                     player.Controller.PrintToCenter($"Map Start ({trigger.Entity.Name})");
 
-                    #if DEBUG
+#if DEBUG
                     player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Green}Map Start Zone");
-                    // player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc -> KeyValues: {trigger.Entity.KeyValues3}");
-                    #endif
+#endif
+                }
+                else
+                {
+                    player.Controller.PrintToChat($"{Config.PluginPrefix} {ChatColors.Yellow}Please try restarting again, replay is still being saved.");
+                }
+            }
+
+            // Stage start zones -- hook into (s)tage#_start
+            else if (Regex.Match(trigger.Entity.Name, "^s([1-9][0-9]?|tage[1-9][0-9]?)_start$").Success)
+            {
+                int stage = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value);
+
+                if (!player.ReplayRecorder.IsRecording)
+                    player.ReplayRecorder.Start();
+
+                player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.STAGE_ZONE_ENTER;
+                player.ReplayRecorder.StageEnterSituations.Add(player.ReplayRecorder.Frames.Count);
+                Console.WriteLine($"STAGE_ZONE_ENTER: player.ReplayRecorder.StageEnterSituations.Add({player.ReplayRecorder.Frames.Count})");
+
+                bool failed_stage = false;
+                if (player.Timer.Stage == stage)
+                    failed_stage = true;
+
+                // Reset/Stop the Stage timer
+                // Save a Stage run when `IsStageMode` is active - (`stage - 1` to get the previous stage data)
+                if (player.Timer.IsStageMode)
+                {
+                    // player.Controller.PrintToChat($"{Config.PluginPrefix} Player ticks higher than 0? {ChatColors.Yellow}{player.Timer.Ticks > 0}");
+                    // player.Controller.PrintToChat($"{Config.PluginPrefix} Player time is faster than StageWR time? {ChatColors.Yellow}{player.Timer.Ticks < CurrentMap.StageWR[stage - 1][style].Ticks}");
+                    // player.Controller.PrintToChat($"{Config.PluginPrefix} No StageWR Exists? {ChatColors.Yellow}{CurrentMap.StageWR[stage - 1][style].ID == -1}");
+                    // player.Controller.PrintToChat($"{Config.PluginPrefix} Not null? {ChatColors.Yellow}{player.Stats.StagePB[stage - 1][style] != null}");
+                    // player.Controller.PrintToChat($"{Config.PluginPrefix} Time faster than existing stage PB? {ChatColors.Yellow}{player.Stats.StagePB[stage - 1][style].Ticks > player.Timer.Ticks}");
+                    if (stage > 1 && !failed_stage && !player.Timer.IsPracticeMode)
+                    {
+                        int stage_run_time = player.Timer.Ticks;
+                        // player.Controller.PrintToChat($"{Config.PluginPrefix} [StageWR (IsStageMode)] Sending to SaveStageTime: {player.Profile.Name}, {stage - 1}, {stage_run_time}");
+                        SaveStageTime(player, stage - 1, stage_run_time);
+                    }
+                    player.Timer.Reset();
+                    player.Timer.IsStageMode = true;
+                    // player.Controller.PrintToChat($"{ChatColors.Red}Resetted{ChatColors.Default} Stage timer for stage {ChatColors.Green}{stage}");
                 }
 
-                // Stage start zones -- hook into (s)tage#_start
-                else if (Regex.Match(trigger.Entity.Name, "^s([1-9][0-9]?|tage[1-9][0-9]?)_start$").Success)
+                player.Timer.Stage = stage;
+
+#if DEBUG
+                Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Timer.IsRunning: {player.Timer.IsRunning}");
+                Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> !player.Timer.IsStageMode: {!player.Timer.IsStageMode}");
+                Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Stats.ThisRun.Checkpoint.Count <= stage: {player.Stats.ThisRun.Checkpoints.Count <= stage}");
+#endif
+
+                // This should patch up re-triggering *player.Stats.ThisRun.Checkpoint.Count < stage*
+                if (player.Timer.IsRunning && !player.Timer.IsStageMode && player.Stats.ThisRun.Checkpoints.Count < stage)
                 {
-                    int stage = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value) - 1;
-                    player.Timer.Stage = stage;
-
-                    #if DEBUG
-                    Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Timer.IsRunning: {player.Timer.IsRunning}");
-                    Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> !player.Timer.IsStageMode: {!player.Timer.IsStageMode}");
-                    Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Stats.ThisRun.Checkpoint.Count <= stage: {player.Stats.ThisRun.Checkpoint.Count <= stage}");
-                    #endif
-
-                    // This should patch up re-triggering *player.Stats.ThisRun.Checkpoint.Count < stage*
-                    if (player.Timer.IsRunning && !player.Timer.IsStageMode && player.Stats.ThisRun.Checkpoint.Count < stage)
+                    // Save Stage MapTime during a Map run
+                    if (stage > 1 && !failed_stage && !player.Timer.IsPracticeMode)
                     {
-                        player.Timer.Checkpoint = stage; // Stage = Checkpoint when in a run on a Staged map
+                        int stage_run_time = player.Timer.Ticks - player.Stats.ThisRun.Ticks; // player.Stats.ThisRun.Ticks should be the Tick we left the previous Stage zone
+                        // player.Controller.PrintToChat($"{Config.PluginPrefix} [StageWR (Map RUN)] Sending to SaveStageTime: {player.Profile.Name}, {stage - 1}, {stage_run_time}");
+                        SaveStageTime(player, stage - 1, stage_run_time);
+                    }
 
-                        #if DEBUG
-                        Console.WriteLine($"============== Initial entity value: {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} | Assigned to `stage`: {Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value) - 1}");
-                        Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Stats.PB[{style}].Checkpoint.Count = {player.Stats.PB[style].Checkpoint.Count}");
-                        #endif
+                    player.Timer.Checkpoint = stage - 1; // Stage = Checkpoint when in a run on a Staged map
 
-                        // Print checkpoint message
-                        player.HUD.DisplayCheckpointMessages(PluginPrefix);
+#if DEBUG
+                    Console.WriteLine($"============== Initial entity value: {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} | Assigned to `stage`: {stage} | player.Timer.Checkpoint: {stage - 1}");
+                    Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Stage start zones) -> player.Stats.PB[{pStyle}].Checkpoint.Count = {player.Stats.PB[pStyle].Checkpoints.Count}");
+#endif
 
-                        // store the checkpoint in the player's current run checkpoints used for Checkpoint functionality
-                        Checkpoint cp2 = new Checkpoint(stage,
+                    // Print checkpoint message
+                    player.HUD.DisplayCheckpointMessages();
+
+                    // store the checkpoint in the player's current run checkpoints used for Checkpoint functionality
+                    if (!player.Stats.ThisRun.Checkpoints.ContainsKey(player.Timer.Checkpoint))
+                    {
+                        Checkpoint cp2 = new Checkpoint(player.Timer.Checkpoint,
                                                         player.Timer.Ticks,
                                                         velocity_x,
                                                         velocity_y,
@@ -175,33 +369,46 @@ public partial class SurfTimer
                                                         -1.0f,
                                                         -1.0f,
                                                         -1.0f,
-                                                        -1.0f,
-                                                        0);
-                        player.Stats.ThisRun.Checkpoint[stage] = cp2;
+                                                        0,
+                                                        1);
+                        player.Stats.ThisRun.Checkpoints[player.Timer.Checkpoint] = cp2;
                     }
-
-                    #if DEBUG
-                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Yellow}Stage {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Start Zone");
-                    #endif
+                    else
+                    {
+                        player.Stats.ThisRun.Checkpoints[player.Timer.Checkpoint].Attempts++;
+                    }
                 }
 
-                // Map checkpoint zones -- hook into map_(c)heck(p)oint#
-                else if (Regex.Match(trigger.Entity.Name, "^map_c(p[1-9][0-9]?|heckpoint[1-9][0-9]?)$").Success)
+#if DEBUG
+                player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.Yellow}Stage {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Start Zone");
+#endif
+            }
+
+            // Map checkpoint zones -- hook into map_(c)heck(p)oint#
+            else if (Regex.Match(trigger.Entity.Name, "^map_c(p[1-9][0-9]?|heckpoint[1-9][0-9]?)$").Success)
+            {
+                int checkpoint = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value);
+                player.Timer.Checkpoint = checkpoint;
+
+                // This should patch up re-triggering *player.Stats.ThisRun.Checkpoint.Count < checkpoint*
+                if (player.Timer.IsRunning && !player.Timer.IsStageMode && player.Stats.ThisRun.Checkpoints.Count < checkpoint)
                 {
-                    int checkpoint = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value);
-                    player.Timer.Checkpoint = checkpoint;
+#if DEBUG
+                    Console.WriteLine($"============== Initial entity value: {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} | Assigned to `checkpoint`: {Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value)}");
+                    Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Checkpoint zones) -> player.Stats.PB[{pStyle}].Checkpoint.Count = {player.Stats.PB[pStyle].Checkpoints.Count}");
+#endif
 
-                    // This should patch up re-triggering *player.Stats.ThisRun.Checkpoint.Count < checkpoint*
-                    if (player.Timer.IsRunning && !player.Timer.IsStageMode && player.Stats.ThisRun.Checkpoint.Count < checkpoint)
+                    if (player.Timer.IsRunning && player.ReplayRecorder.IsRecording)
                     {
-                        #if DEBUG
-                        Console.WriteLine($"============== Initial entity value: {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} | Assigned to `checkpoint`: {Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value) - 1}");
-                        Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Checkpoint zones) -> player.Stats.PB[{style}].Checkpoint.Count = {player.Stats.PB[style].Checkpoint.Count}");
-                        #endif
-                        
-                        // Print checkpoint message
-                        player.HUD.DisplayCheckpointMessages(PluginPrefix);
+                        player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.CHECKPOINT_ZONE_ENTER;
+                        player.ReplayRecorder.CheckpointEnterSituations.Add(player.Timer.Ticks);
+                    }
 
+                    // Print checkpoint message
+                    player.HUD.DisplayCheckpointMessages();
+
+                    if (!player.Stats.ThisRun.Checkpoints.ContainsKey(checkpoint))
+                    {
                         // store the checkpoint in the player's current run checkpoints used for Checkpoint functionality
                         Checkpoint cp2 = new Checkpoint(checkpoint,
                                                         player.Timer.Ticks,
@@ -211,18 +418,183 @@ public partial class SurfTimer
                                                         -1.0f,
                                                         -1.0f,
                                                         -1.0f,
-                                                        -1.0f,
-                                                        0);
-                        player.Stats.ThisRun.Checkpoint[checkpoint] = cp2;
+                                                        0,
+                                                        1);
+                        player.Stats.ThisRun.Checkpoints[checkpoint] = cp2;
                     }
-
-                    #if DEBUG
-                    player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.LightBlue}Checkpoint {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Zone");
-                    #endif
+                    else
+                    {
+                        player.Stats.ThisRun.Checkpoints[checkpoint].Attempts++;
+                    }
                 }
+
+#if DEBUG
+                player.Controller.PrintToChat($"CS2 Surf DEBUG >> CBaseTrigger_{ChatColors.Lime}StartTouchFunc{ChatColors.Default} -> {ChatColors.LightBlue}Checkpoint {Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value} Zone");
+#endif
             }
 
-            return HookResult.Continue;
+            // Bonus start zones -- hook into (b)onus#_start
+            else if (Regex.Match(trigger.Entity.Name, "^b([1-9][0-9]?|onus[1-9][0-9]?)_start$").Success)
+            {            
+                int bonus = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value);
+                player.Timer.Bonus = bonus;
+
+                player.Timer.Reset();
+                player.Timer.IsBonusMode = true;
+
+
+                player.ReplayRecorder.Reset();
+                player.ReplayRecorder.Start(); // Start replay recording
+                player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.START_ZONE_ENTER;
+                player.ReplayRecorder.BonusSituations.Add(player.ReplayRecorder.Frames.Count);
+                Console.WriteLine($"START_ZONE_ENTER: player.ReplayRecorder.BonusSituations.Add({player.ReplayRecorder.Frames.Count})");
+
+                player.Controller.PrintToCenter($"Bonus Start ({trigger.Entity.Name})");
+
+#if DEBUG
+                Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Bonus start zones) -> player.Timer.IsRunning: {player.Timer.IsRunning}");
+                Console.WriteLine($"CS2 Surf DEBUG >> CBaseTrigger_StartTouchFunc (Bonus start zones) -> !player.Timer.IsBonusMode: {!player.Timer.IsBonusMode}");
+#endif
+            }
+
+            // Bonus end zones -- hook into (b)onus#_end
+            else if (Regex.Match(trigger.Entity.Name, "^b([1-9][0-9]?|onus[1-9][0-9]?)_end$").Success && player.Timer.IsBonusMode && player.Timer.IsRunning)
+            {
+                // To-do: verify the bonus trigger being hit!
+                int bonus_idx = Int32.Parse(Regex.Match(trigger.Entity.Name, "[0-9][0-9]?").Value);
+
+                player.Timer.Stop();
+                player.ReplayRecorder.CurrentSituation = ReplayFrameSituation.END_ZONE_ENTER;
+                player.ReplayRecorder.BonusSituations.Add(player.Timer.Ticks);
+
+                player.Stats.ThisRun.Ticks = player.Timer.Ticks; // End time for the run
+                player.Stats.ThisRun.EndVelX = velocity_x; // End pre speed for the run
+                player.Stats.ThisRun.EndVelY = velocity_y; // End pre speed for the run
+                player.Stats.ThisRun.EndVelZ = velocity_z; // End pre speed for the run
+
+                bool saveBonusTime = false;
+                string PracticeString = "";
+                if (player.Timer.IsPracticeMode)
+                    PracticeString = $"({ChatColors.Grey}Practice{ChatColors.Default}) ";
+
+                if (player.Timer.Ticks < CurrentMap.BonusWR[bonus_idx][pStyle].Ticks) // Player beat the Bonus WR
+                {
+                    saveBonusTime = true;
+                    int timeImprove = CurrentMap.BonusWR[bonus_idx][pStyle].Ticks - player.Timer.Ticks;
+                    Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} has set a new {ChatColors.Yellow}Bonus {bonus_idx}{ChatColors.Default} record with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, beating the old record by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(CurrentMap.BonusWR[bonus_idx][pStyle].Ticks)}{ChatColors.Default})");
+                }
+                else if (CurrentMap.BonusWR[bonus_idx][pStyle].ID == -1) // No Bonus record was set on the map
+                {
+                    saveBonusTime = true;
+                    Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} set the first {ChatColors.Yellow}Bonus {bonus_idx}{ChatColors.Default} record at {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}!");
+                }
+                else if (player.Stats.BonusPB[bonus_idx][pStyle].Ticks <= 0) // Player first ever PersonalBest for the bonus
+                {
+                    saveBonusTime = true;
+                    player.Controller.PrintToChat($"{Config.PluginPrefix} {PracticeString}You finished the {ChatColors.Yellow}Bonus {bonus_idx}{ChatColors.Default} in {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, setting your new Personal Best!");
+                }
+                else if (player.Timer.Ticks < player.Stats.BonusPB[bonus_idx][pStyle].Ticks) // Player beating their existing PersonalBest for the bonus
+                {
+                    saveBonusTime = true;
+                    int timeImprove = player.Stats.BonusPB[bonus_idx][pStyle].Ticks - player.Timer.Ticks;
+                    Server.PrintToChatAll($"{Config.PluginPrefix} {PracticeString}{ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} beat their {ChatColors.Yellow}Bonus {bonus_idx}{ChatColors.Default} Personal Best with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}, improving by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(player.Stats.PB[pStyle].Ticks)}{ChatColors.Default})");
+                }
+                else // Player did not beat their existing personal best for the bonus
+                {
+                    player.Controller.PrintToChat($"{Config.PluginPrefix} {PracticeString}You finished {ChatColors.Yellow}Bonus {bonus_idx}{ChatColors.Default} in {ChatColors.Yellow}{PlayerHUD.FormatTime(player.Timer.Ticks)}{ChatColors.Default}!");
+                }
+
+                // To-do: save to DB
+                if (!player.Timer.IsPracticeMode)
+                {
+                    /*
+                    AddTimer(1.5f, () =>
+                    {
+                        API_CurrentRun bonus_time = new API_CurrentRun
+                        {
+                            player_id = player.Profile.ID,
+                            map_id = player.CurrMap.ID,
+                            style = pStyle,
+                            type = 1,
+                            stage = bonus_idx,
+                            run_time = player.Stats.ThisRun.Ticks,
+                            run_date = (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds(),
+                            replay_frames = player.ReplayRecorder.SerializeReplay()
+                        };
+
+                        Task.Run(async () =>
+                        {
+                            await ApiMethod.POST("/surftimer/savebonustime", bonus_time);
+                            player.Stats.LoadBonusTime(player);
+                            await CurrentMap.Get_Map_Record_Runs(); // Reload the Map record and totals for the HUD
+                                                                    // await CurrentMap.ApiGetMapRecordAndTotals(); // Reload the Map record and totals for the HUD
+                        });
+                    });
+                    */
+                    if (saveBonusTime)
+                    {
+                        player.ReplayRecorder.IsSaving = true;
+                        AddTimer(1.0f, async () =>
+                        {
+                            await player.Stats.ThisRun.SaveMapTime(player, bonus: bonus_idx); // Save the Bonus MapTime data
+                        });
+                    }
+                }
+            }
+        }
+        return HookResult.Continue;
+    }
+
+    /// <summary>
+    /// Deals with saving a Stage MapTime (Type 2) in the Database.
+    /// Should deal with `IsStageMode` runs, Stages during Map Runs and also Last Stage.
+    /// </summary>
+    /// <param name="player">Player object</param>
+    /// <param name="stage">Stage to save</param>
+    /// <param name="saveLastStage">Is it the last stage?</param>
+    /// <param name="stage_run_time">Run Time (Ticks) for the stage run</param>
+    void SaveStageTime(Player player, int stage = -1, int stage_run_time = -1, bool saveLastStage = false)
+    {
+        // player.Controller.PrintToChat($"{Config.PluginPrefix} SaveStageTime received: {player.Profile.Name}, {stage}, {stage_run_time}, {saveLastStage}");
+        int pStyle = player.Timer.Style;
+        if (
+            stage_run_time < CurrentMap.StageWR[stage][pStyle].Ticks ||
+            CurrentMap.StageWR[stage][pStyle].ID == -1 ||
+            player.Stats.StagePB[stage][pStyle] != null && player.Stats.StagePB[stage][pStyle].Ticks > stage_run_time
+        )
+        {
+            if (stage_run_time < CurrentMap.StageWR[stage][pStyle].Ticks) // Player beat the Stage WR
+            {
+                int timeImprove = CurrentMap.StageWR[stage][pStyle].Ticks - stage_run_time;
+                Server.PrintToChatAll($"{Config.PluginPrefix} {ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} has set a new {ChatColors.Yellow}Stage {stage}{ChatColors.Default} record with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(stage_run_time)}{ChatColors.Default}, beating the old record by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(CurrentMap.StageWR[stage][pStyle].Ticks)}{ChatColors.Default})");
+            }
+            else if (CurrentMap.StageWR[stage][pStyle].ID == -1) // No Stage record was set on the map
+            {
+                Server.PrintToChatAll($"{Config.PluginPrefix} {ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} set the first {ChatColors.Yellow}Stage {stage}{ChatColors.Default} record at {ChatColors.Gold}{PlayerHUD.FormatTime(stage_run_time)}{ChatColors.Default}!");
+            }
+            else if (player.Stats.StagePB[stage][pStyle] != null && player.Stats.StagePB[stage][pStyle].ID == -1) // Player first Stage personal best
+            {
+                player.Controller.PrintToChat($"{Config.PluginPrefix} You finished {ChatColors.Yellow}Stage {stage}{ChatColors.Default} in {ChatColors.Gold}{PlayerHUD.FormatTime(stage_run_time)}{ChatColors.Default}, setting your new Personal Best!");
+            }
+            else if (player.Stats.StagePB[stage][pStyle] != null && player.Stats.StagePB[stage][pStyle].Ticks > stage_run_time) // Player beating their existing Stage personal best
+            {
+                int timeImprove = player.Stats.StagePB[stage][pStyle].Ticks - stage_run_time;
+                Server.PrintToChatAll($"{Config.PluginPrefix} {ChatColors.Lime}{player.Controller.PlayerName}{ChatColors.Default} beat their {ChatColors.Yellow}Stage {stage}{ChatColors.Default} Personal Best with a time of {ChatColors.Gold}{PlayerHUD.FormatTime(stage_run_time)}{ChatColors.Default}, improving by {ChatColors.Green}-{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default}! (Previous: {ChatColors.BlueGrey}{PlayerHUD.FormatTime(player.Stats.StagePB[stage][pStyle].Ticks)}{ChatColors.Default})");
+            }
+
+            player.ReplayRecorder.IsSaving = true;
+            AddTimer(1.0f, async () =>
+            {
+                // Save stage run
+                Console.WriteLine($"==== OnTriggerStartTouch -> SaveStageTime -> [StageWR (IsStageMode? {player.Timer.IsStageMode} | Last? {saveLastStage})] Saving Stage {stage} ({stage}) time of {PlayerHUD.FormatTime(stage_run_time)} ({stage_run_time})");
+                await player.Stats.ThisRun.SaveMapTime(player, stage: stage, run_ticks: stage_run_time); // Save the Stage MapTime PB data
+            });
+        }
+        else if (stage_run_time > CurrentMap.StageWR[stage][pStyle].Ticks && player.Timer.IsStageMode) // Player is behind the Stage WR for the map
+        {
+            int timeImprove = stage_run_time - CurrentMap.StageWR[stage][pStyle].Ticks;
+            player.Controller.PrintToChat($"{Config.PluginPrefix} You are behind the {ChatColors.Yellow}Stage {stage}{ChatColors.Default} record with a time of {ChatColors.Grey}{PlayerHUD.FormatTime(stage_run_time)}{ChatColors.Default}, missing the record by {ChatColors.Red}+{PlayerHUD.FormatTime(timeImprove)}{ChatColors.Default} ({ChatColors.Gold}{PlayerHUD.FormatTime(CurrentMap.StageWR[stage][pStyle].Ticks)}{ChatColors.Default})!");
         }
     }
 }
+
