@@ -5,7 +5,7 @@ namespace SurfTimer;
 /// </summary>
 internal class CurrentRun
 {
-    public Dictionary<int, Checkpoint> Checkpoint { get; set; } // Current RUN checkpoints tracker
+    public Dictionary<int, Checkpoint> Checkpoints { get; set; } // Current RUN checkpoints tracker
     public int Ticks { get; set; } // To-do: will be the last (any) zone end touch time
     public float StartVelX { get; set; } // This will store MAP START VELOCITY X
     public float StartVelY { get; set; } // This will store MAP START VELOCITY Y
@@ -19,7 +19,7 @@ internal class CurrentRun
     // Constructor
     public CurrentRun()
     {
-        Checkpoint = new Dictionary<int, Checkpoint>();
+        Checkpoints = new Dictionary<int, Checkpoint>();
         Ticks = 0;
         StartVelX = 0.0f;
         StartVelY = 0.0f;
@@ -32,7 +32,7 @@ internal class CurrentRun
 
     public void Reset()
     {
-        Checkpoint.Clear();
+        Checkpoints.Clear();
         Ticks = 0;
         StartVelX = 0.0f;
         StartVelY = 0.0f;
@@ -45,46 +45,152 @@ internal class CurrentRun
     }
 
     /// <summary>
-    /// Saves the player's run to the database and reloads the data for the player.
-    /// NOTE: Not re-loading any data at this point as we need `LoadMapTimesData` to be called from here as well, otherwise we may not have the `this.ID` populated
+    /// Saves the player's run to the database. 
     /// </summary>
-    public void SaveMapTime(Player player, TimerDatabase DB)
+    /// <param name="player">Player object</param>
+    /// <param name="bonus">Bonus number</param>
+    /// <param name="stage">Stage number</param>
+    /// <param name="run_ticks">Ticks for the run - used for Stage and Bonus entries</param>
+    public async Task SaveMapTime(Player player, int bonus = 0, int stage = 0, int run_ticks = -1)
     {
         // Add entry in DB for the run
-        // To-do: add `type`
-        int style = player.Timer.Style;
-        string replay_frames = player.ReplayRecorder.SerializeReplay();
-        Task<int> updatePlayerRunTask = DB.Write($@"
-            INSERT INTO `MapTimes` 
-            (`player_id`, `map_id`, `style`, `type`, `stage`, `run_time`, `start_vel_x`, `start_vel_y`, `start_vel_z`, `end_vel_x`, `end_vel_y`, `end_vel_z`, `run_date`, `replay_frames`) 
-            VALUES ({player.Profile.ID}, {player.CurrMap.ID}, {style}, 0, 0, {player.Stats.ThisRun.Ticks}, 
-            {player.Stats.ThisRun.StartVelX}, {player.Stats.ThisRun.StartVelY}, {player.Stats.ThisRun.StartVelZ}, {player.Stats.ThisRun.EndVelX}, {player.Stats.ThisRun.EndVelY}, {player.Stats.ThisRun.EndVelZ}, {(int)DateTimeOffset.UtcNow.ToUnixTimeSeconds()}, '{replay_frames}') 
-            ON DUPLICATE KEY UPDATE run_time=VALUES(run_time), start_vel_x=VALUES(start_vel_x), start_vel_y=VALUES(start_vel_y), 
-            start_vel_z=VALUES(start_vel_z), end_vel_x=VALUES(end_vel_x), end_vel_y=VALUES(end_vel_y), end_vel_z=VALUES(end_vel_z), run_date=VALUES(run_date), replay_frames=VALUES(replay_frames);
-        ");
-        if (updatePlayerRunTask.Result <= 0)
-            throw new Exception($"CS2 Surf ERROR >> internal class PersonalBest -> SaveMapTime -> Failed to insert/update player run in database. Player: {player.Profile.Name} ({player.Profile.SteamID})");
-        updatePlayerRunTask.Dispose();
+        // PrintSituations(player);
+        string replay_frames = player.ReplayRecorder.TrimReplay(player, stage != 0 ? 2 : bonus != 0 ? 1 : 0, stage == SurfTimer.CurrentMap.Stages);
 
-        // Will have to LoadMapTimesData right here as well to get the ID of the run we just inserted
-        // this.SaveCurrentRunCheckpoints(player, DB); // Save checkpoints for this run
-        // this.LoadCheckpointsForRun(DB); // Re-Load checkpoints for this run
+        Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task SaveMapTime -> Sending total of {replay_frames.Length} replay frames");
+        if (Config.API.GetApiOnly())
+        {
+            return;
+        }
+        else
+        {
+            await InsertMapTime(player, bonus, stage, run_ticks, replay_frames, true);
+
+            if (stage != 0 || bonus != 0)
+            {
+                Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task SaveMapTime -> Inserted an entry for {(stage != 0 ? "Stage" : "Bonus")} {(stage != 0 ? stage : bonus)} - {run_ticks}");
+            }
+            else
+            {
+                await SaveCurrentRunCheckpoints(player, true); // Save this run's checkpoints
+            }
+
+            await player.CurrMap.Get_Map_Record_Runs(); // Reload the times for the Map
+        }
+    }
+
+    public void PrintSituations(Player player)
+    {
+        Console.WriteLine($"========================== FOUND SITUATIONS ==========================");
+        for (int i = 0; i < player.ReplayRecorder.Frames.Count; i++)
+        {
+            ReplayFrame x = player.ReplayRecorder.Frames[i];
+            switch (x.Situation)
+            {
+                case ReplayFrameSituation.START_ZONE_ENTER:
+                    Console.WriteLine($"START_ZONE_ENTER: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.START_ZONE_EXIT:
+                    Console.WriteLine($"START_ZONE_EXIT: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.STAGE_ZONE_ENTER:
+                    Console.WriteLine($"STAGE_ZONE_ENTER: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.STAGE_ZONE_EXIT:
+                    Console.WriteLine($"STAGE_ZONE_EXIT: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.CHECKPOINT_ZONE_ENTER:
+                    Console.WriteLine($"CHECKPOINT_ZONE_ENTER: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.CHECKPOINT_ZONE_EXIT:
+                    Console.WriteLine($"CHECKPOINT_ZONE_EXIT: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.END_ZONE_ENTER:
+                    Console.WriteLine($"END_ZONE_ENTER: {i} | Situation {x.Situation}");
+                    break;
+                case ReplayFrameSituation.END_ZONE_EXIT:
+                    Console.WriteLine($"END_ZONE_EXIT: {i} | Situation {x.Situation}");
+                    break;
+            }
+        }
+        Console.WriteLine($"==========================                  ==========================");
+    }
+
+    /// <summary>
+    /// Saves the CurrentRun of the player to the database. Does NOT support Bonus entries yet.
+    /// </summary>
+    /// <param name="player">Player object</param>
+    /// <param name="bonus">Bonus number</param>
+    /// <param name="stage">Stage number</param>
+    /// <param name="run_ticks">Ticks for the run</param>
+    /// <param name="replay_frames">Replay frames</param>
+    /// <param name="reloadData">Whether to reload the PersonalBest data for the Player.</param>
+    /// <returns></returns>
+    public async Task InsertMapTime(Player player, int bonus = 0, int stage = 0, int run_ticks = -1, string replay_frames = "", bool reloadData = false)
+    {
+        int playerId = player.Profile.ID;
+        int mapId = player.CurrMap.ID;
+        int style = player.Timer.Style;
+        int ticks = run_ticks == -1 ? player.Stats.ThisRun.Ticks : run_ticks;
+        // int ticks = player.Stats.ThisRun.Ticks;
+        int type = stage != 0 ? 2 : bonus != 0 ? 1 : 0;
+        float startVelX = player.Stats.ThisRun.StartVelX;
+        float startVelY = player.Stats.ThisRun.StartVelY;
+        float startVelZ = player.Stats.ThisRun.StartVelZ;
+        float endVelX = player.Stats.ThisRun.EndVelX;
+        float endVelY = player.Stats.ThisRun.EndVelY;
+        float endVelZ = player.Stats.ThisRun.EndVelZ;
+
+        if (Config.API.GetApiOnly()) // API Calls
+        {
+            // API Insert map goes here
+        }
+        else // MySQL Queries
+        {
+            int updatePlayerRunTask = await SurfTimer.DB.WriteAsync(
+                string.Format(Config.MySQL.Queries.DB_QUERY_CR_INSERT_TIME, playerId, mapId, style, type, type == 2 ? stage : type == 1 ? bonus : 0, ticks, startVelX, startVelY, startVelZ, endVelX, endVelY, endVelZ, (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds(), replay_frames));
+            if (updatePlayerRunTask <= 0)
+            {
+                Exception ex = new($"CS2 Surf ERROR >> internal class CurrentRun -> public async Task InsertMapTime -> Failed to insert/update player run in database. Player: {player.Profile.Name} ({player.Profile.SteamID})");
+                throw ex;
+            }
+
+            if (reloadData && type == 0)
+            {
+                Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task InsertMapTime -> Will reload MapTime (Type {type}) data for '{player.Profile.Name}' (ID {player.Stats.PB[player.Timer.Style].ID}))");
+                await player.Stats.PB[style].PB_LoadPlayerSpecificMapTimeData(player); // Load the Map MapTime PB data again (will refresh the MapTime ID for the Checkpoints query)
+            }
+            else if (reloadData && type == 1)
+            {
+                Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task InsertMapTime -> Will reload Bonus MapTime (Type {type}) data for '{player.Profile.Name}' (ID {player.Stats.BonusPB[bonus][style].ID}))");
+                await player.Stats.BonusPB[bonus][style].PB_LoadPlayerSpecificMapTimeData(player); // Load the Bonus MapTime PB data again (will refresh the MapTime ID)
+            }
+            else if (reloadData && type == 2)
+            {
+                Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task InsertMapTime -> Will reload Stage MapTime (Type {type}) data for '{player.Profile.Name}' (ID {player.Stats.StagePB[stage][style].ID}))");
+                await player.Stats.StagePB[stage][style].PB_LoadPlayerSpecificMapTimeData(player); // Load the Stage MapTime PB data again (will refresh the MapTime ID)
+            }
+        }
     }
 
     /// <summary>
     /// Saves the `CurrentRunCheckpoints` dictionary to the database
-    /// We need the correct `this.ID` to be populated before calling this method otherwise Query will fail
     /// </summary>
-    public async Task SaveCurrentRunCheckpoints(Player player, TimerDatabase DB)
+    /// <param name="player">Player object</param>
+    /// <param name="reloadData">Whether to reload the PersonalBest Checkpoints data for the Player.</param>
+    public async Task SaveCurrentRunCheckpoints(Player player, bool reloadData = false)
     {
+        Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> SaveCurrentRunCheckpoints -> Will send {player.Stats.ThisRun.Checkpoints.Count} ({this.Checkpoints.Count}) checkpoints to DB....");
         int style = player.Timer.Style;
+        int mapTimeId = player.Stats.PB[style].ID;
         List<string> commands = new List<string>();
         // Loop through the checkpoints and insert/update them in the database for the run
-        foreach (var item in player.Stats.ThisRun.Checkpoint)
+        // foreach (var item in player.Stats.ThisRun.Checkpoints)
+        foreach (var item in this.Checkpoints)
         {
             int cp = item.Key;
             int ticks = item.Value!.Ticks;
-            int runTime = item.Value!.Ticks / 64; // Runtime in decimal
+            int endTouch = item.Value!.EndTouch;
             double startVelX = item.Value!.StartVelX;
             double startVelY = item.Value!.StartVelY;
             double startVelZ = item.Value!.StartVelZ;
@@ -93,35 +199,33 @@ internal class CurrentRun
             double endVelZ = item.Value!.EndVelZ;
             int attempts = item.Value!.Attempts;
 
-            #if DEBUG
-            Console.WriteLine($"CP: {cp} | MapTime ID: {player.Stats.PB[style].ID} | Time: {runTime} | Ticks: {ticks} | startVelX: {startVelX} | startVelY: {startVelY} | startVelZ: {startVelZ} | endVelX: {endVelX} | endVelY: {endVelY} | endVelZ: {endVelZ}");
-            Console.WriteLine($@"CS2 Surf DEBUG >> internal class Checkpoint : PersonalBest -> SaveCurrentRunCheckpoints -> 
-                INSERT INTO `Checkpoints` 
-                (`maptime_id`, `cp`, `run_time`, `start_vel_x`, `start_vel_y`, `start_vel_z`, 
-                `end_vel_x`, `end_vel_y`, `end_vel_z`, `attempts`, `end_touch`) 
-                VALUES ({player.Stats.PB[style].ID}, {cp}, {runTime}, {startVelX}, {startVelY}, {startVelZ}, {endVelX}, {endVelY}, {endVelZ}, {attempts}, {ticks}) ON DUPLICATE KEY UPDATE 
-                run_time=VALUES(run_time), start_vel_x=VALUES(start_vel_x), start_vel_y=VALUES(start_vel_y), start_vel_z=VALUES(start_vel_z), 
-                end_vel_x=VALUES(end_vel_x), end_vel_y=VALUES(end_vel_y), end_vel_z=VALUES(end_vel_z), attempts=VALUES(attempts), end_touch=VALUES(end_touch);
+#if DEBUG
+            Console.WriteLine($"CP: {cp} | MapTime ID: {mapTimeId} | Time: {endTouch} | Ticks: {ticks} | startVelX: {startVelX} | startVelY: {startVelY} | startVelZ: {startVelZ} | endVelX: {endVelX} | endVelY: {endVelY} | endVelZ: {endVelZ}");
+            Console.WriteLine($@"CS2 Surf DEBUG >> internal class CurrentRun -> SaveCurrentRunCheckpoints -> 
+                {string.Format(
+                    Config.MySQL.Queries.DB_QUERY_CR_INSERT_CP,
+                    mapTimeId, cp, ticks, startVelX, startVelY, startVelZ, endVelX, endVelY, endVelZ, attempts, endTouch)}
             ");
-            #endif
+#endif
 
             // Insert/Update CPs to database
             // Check if the player has PB object initialized and if the player's character is currently active in the game
             if (item.Value != null && player.Controller.PlayerPawn.Value != null)
             {
-                string command = $@"
-                    INSERT INTO `Checkpoints` 
-                    (`maptime_id`, `cp`, `run_time`, `start_vel_x`, `start_vel_y`, `start_vel_z`, 
-                    `end_vel_x`, `end_vel_y`, `end_vel_z`, `attempts`, `end_touch`) 
-                    VALUES ({player.Stats.PB[style].ID}, {cp}, {runTime}, {startVelX}, {startVelY}, {startVelZ}, {endVelX}, {endVelY}, {endVelZ}, {attempts}, {ticks}) 
-                    ON DUPLICATE KEY UPDATE 
-                    run_time=VALUES(run_time), start_vel_x=VALUES(start_vel_x), start_vel_y=VALUES(start_vel_y), start_vel_z=VALUES(start_vel_z), 
-                    end_vel_x=VALUES(end_vel_x), end_vel_y=VALUES(end_vel_y), end_vel_z=VALUES(end_vel_z), attempts=VALUES(attempts), end_touch=VALUES(end_touch);
-                ";
+                string command = string.Format(
+                    Config.MySQL.Queries.DB_QUERY_CR_INSERT_CP,
+                    mapTimeId, cp, ticks, startVelX, startVelY, startVelZ, endVelX, endVelY, endVelZ, attempts, endTouch
+                );
                 commands.Add(command);
             }
         }
-        await DB.Transaction(commands);
-        player.Stats.ThisRun.Checkpoint.Clear();
+        await SurfTimer.DB.TransactionAsync(commands);
+        player.Stats.ThisRun.Checkpoints.Clear();
+
+        if (reloadData)
+        {
+            Console.WriteLine($"CS2 Surf DEBUG >> internal class CurrentRun -> public async Task SaveCurrentRunCheckpoints -> Will reload Checkpoints data for {player.Profile.Name} (ID {player.Stats.PB[player.Timer.Style].ID})");
+            await player.Stats.PB[player.Timer.Style].PB_LoadCheckpointsData(); // Load the Checkpoints data again
+        }
     }
 }
